Multiple changes were made.  Most of the changes were to the code Sean wrote in order to integrate it.  However, one unrelated change was performed on the Crawler class.
The Browser class now automatically populates the Start and End properties of all generated HttpRequestResults.

One major change was made to the Crawler class, and as a result to several files in the CrawlerTestApp.  I changed the purpose of the WasRetrieved property for all HttpRequestResults to reflect if the Crawler had retrieved it.  The WasRetrieved method will now be false for all HttpRequestResults whose ResultUrl property is not contained in the set of hosts retrieved from the seeds list.  I changed the AbsoluteUrl tests in the CrawlerTestApp to use the IP address instead of the host returned from an aspx method call.

I made one change to the Link model.  I added an additional property, IsBroken.  The purpose of that property is described in the code comments.

I made two changes to the TestResultsFactory.  I made it so the OutputTypeErrorInfo GenerateErrorInfo(HttpRequestResult result) method created an appropriate message given the state of the HttpRequestResult object passed into it.  I also added a function bool IsSuccess(HttpRequestResult result) to singularize the code used to determine if an HttpRequestResult should show up as a failed test case or not.