Multiple changes were made to the configuration sections.  An optional network credentials element was added for basic and digest authentication needs.  The browsers section was made to not be required as the default constructor creates a collection with one browser.  Three attributes were added to the link checker configuration element; a name, description, and file to store the test results.  An explicit conversion operator was added to the seeds collection to allow a convenience constructor to be added to the Crawler class.
Three changes were performed to the Browser class.  The first being that multiple constructors were added to allow a browser object to be created from their respective configuration classes.  The second change causes the newly added BrowserUsed member of the HttpRequestResult objects to be populated.  The final change allows the Browser class to use authentication credentials to access object protected resources on a server.

Two changes were made to the HttpRequestResult class.  A member named BrowserUsed of the Browser class was added.  In addition, an implementation of the IComparable interface was added to allow objects of the class to be stored in a SortedSet.

Three changes were made to the TestResultsFactory class.  The unit test names now include the browser name, and the request URL.  No error info element is generated for unit tests that pass.  The IsSuccess method was corrected to return false when the Error member of the HttpRequestResult parameter was not null instead of when it was null.

Along with multiple changes in the previously mentioned classes, a number of classes had just a minor change.  The WebSiteTestSuiteGenerator class was completed.  A convenience constructor was added to the Crawler class.  The program class was changed to call the newly created GenerateTests method of the WebSiteTestSuiteGenerator class.  Another unit test was added to the set of Uri tests for hash tags.