Refs LOD-67.
Two additions and two changes were performed.  The Source property of the Image class was removed as it is not needed to contain information.  All of the regexs of the HtmlParser class was modified to use a back-reference instead of a character class to find the closing quotation mark for their respective attributes.  The ParserTestParts class was added to decouple the test case data originally coupled with the TestHyperLink class.  Three new test classes were added to assist in testing the HtmlParser's regular expressions.