Refs LOD-61 and LOD-65
The namespace for the BrowserCollection, BrowserConfigElement, LinkCheckerConfigSection, NetworkCredentialsElement, SeedCollection, and SeedConfigElement were changed to IInteractive.WebConsole.  A try-catch was added to the WebSiteTestSuiteGenerator object defined in the Program class' Execute method for seed URIs that cause UriFormatExceptions.

The new classes under IInteractive.WebTest/TestConfigurationClasses directory, and the TestExperimentalConfigs test class were added for testing a design assumption.  The design rests on the assumption that a Parent property can be added to each of the custom Configuration classes in WebConsole.  Past tests have shown that other methods may not be feasible.  This is important for retrieving the the max number of automatic redirects for remote sites.

Tests cases were added to test that the new design will work.  Currently, the tests do not compile.  These tests were added to the TestBrowser and TestConfigurationParenting classes.  Currently, the TestConfigurationParenting class tests both the attribute to be added with LOD-61 and LOD-65.  The TestBrowser currently only tests the attribute to be added with LOD-61.

Text templates are now made available to the WebTest project through resources.  The Resources.resx and Resources1.Designer.cs files make the files in the IInteractive.WebTest/Resources/ directory available during unit test execution.