Refs LOD-61
An extension to the range of valid values was made for the LinkCheckerConfigSection's MaxRemoteAutomaticRedirects attribute.  This change was performed to fix multiple failing regression tests caused by the DefaultValue populating the property.

The Crawler class was evolved to include an IsRemote method to properly call the Browser class' Get method.  This change was made in response to multiple failing regression tests caused by the exception thrown by the previous implementation.