Refs LOD-61 and LOD-65
A test web page was added to the CrawlerTestApp for testing timeouts.  The BrowserCollection class was modified to instantiate all Parent properties of element retrieved through the indexer operations.  The LinkCheckerConfigSetcion was modified to instantiate the Parent property of all BrowserCollection objects retrieved through its Browsers member.  The BrowserConfigElement class was modified to retrieved the parent property values for Timeout and MaxRemoteAutomaticRedirects.  The timeout attribute of the CaseE resource was modified for two new test cases added to the TestBrowser class.  Two test cases were added to the TestBrowser class to ensure that the timeout member was being used correctly.  The TestBrowsersParentingCaseC test case apart of the TestConfigurationParenting class was corrected.  Two methods for retrieving the remote test site were added to the TestCrawler class.