Refs LOD-63.
Multiple changes were made to incorporate the tests for the forbidden configuration items.  The new test cases are spread across three test classes.  New test dependencies were added to the CrawlerTestApp and the resources of the WebTest project.

An assumption on the CallbackValidator was pushed into the set of experimental config tests.  A member was added to the ConfElem class with a CallbackValidator attribute.  The TestExperimentalConfigs test class was modified in order to showcase the behavior of that attribute.

The TestCrawler class was modified to abstract out the local and remote URLs used to access the CrawlerTestApp.

A test case was added to the TestProgram class to make sure that a ConfigurationErrorsException is thrown when a seed's host is found to match a forbidden host.  A test class was added, TestForbiddenConfiguration, in order to determine that the forbidden collection is not required.  A test case with an overlapping purpose as the one added to TestProgram was also included in the TestForbiddenConfiguration class.

Multiple testing resources were added.  A new directory was added to the CrawlerTestApp project with accessible resources.  Three new configuration files were added to the resources for the WebTest project.