Refs LOD-64.
Added the test cases and their resources.  The BrokenLinksTemplate.txt was added to give a template for the configuration file.  The Resources.resx and Resources1.Designer were modified to make the BrokenLinksTemplate.txt accessible by unit test code.  A BrokenLinkConfigTests directory was added to the CrawlerTestApp project for these tests.

At current these new tests do not compile as test driven development is being followed.

The unit tests specified by the TestProgram class was increased by three.  Two for program executions that are expected to an error.  One for program executions that are expected to return successfully.

The unit tests specified by the TestWebSiteTestSuiteGenerator class was increased by two.  A template method was created.  Two test methods that use that template were added.