Refs LOD-76
Changed the class level accessibility modifier for the following classes.
# Error
# Errors
# Fault
# Faults
# Warning
# Warnings
# WarningPotentialIssues

The TestHtmlValidate test class was renamed to TestHtmlValidator.  A configuration template was added to the WebTest project's resources for the TestHtmlValidator test cases.  Four test cases were added to the TestHtmlValidator test class.  Five new resources were added to the CrawlerTestApp for testing purposes.