Refs LOD-76.
Moved the MarkupValidator project into the WebConsole project.

Changed the constructor for the WebSiteTestSuiteGenerator to include two bool arguments.  One determines whether the HTML validation should be run and the other determines if the CSS validation should be run.  Two members were added to the class to store these values.  The Program and TestWebSiteTestSuiteGenerator were modified to use the new constructor.

An HtmlValidationResult object was added to the HttpRequestResult class to store the result of the W3C validation.

The HTMLValidate class was split into two separate classes.  The HtmlValidator class encapsulates the data and methods for validating a retrieved HttpRequestResult object.  The HtmlValidationResult encapsulates the information retrieved from the W3C validation service.